feat: Changed uint size for fee and lastUpdated for gas checks

+++ src/libraries/Events.sol
@@ -3,7 +3,7 @@
 pragma solidity 0.8.10;
 
 library Events {
-    event FeeUpdated(uint256 oldFee, uint256 newFee);
+    event FeeUpdated(uint64 oldFee, uint64 newFee);
     event FeeTaken(uint256 shares);
     event FeesWithdrawn(address indexed to, uint256 amount);
     event AavePoolUpdated(address newAavePool);
